refactor(dashboards): share lane-timeline base between state-timeline and status-history
Extract collectLaneSamples/assignStateColors and a shared LaneTimelineChart
(brush, axis, legend, tooltip) used by both StateTimeline and StatusHistory,
which previously duplicated ~90% of their data builder and renderer. The
shared tooltip uses SeriesTooltipContent. Each viz keeps only its distinct
item construction (hold-until-next segments vs fixed-width cells).
